Refactor to move all current setting logic to ensureCurrentUser
