+++ app/javascript/components/about_page/AboutPageContainer.jsx
@@ -2,21 +2,16 @@ import React from 'react'
 import HeaderContainer from "../nav/HeaderContainer"
 import FooterContainer from "../nav/FooterContainer"
 import AboutPage from "./AboutPage"
+import ensureCurrentUser from "../../apps/ensureCurrentUser";
 
-export default class AboutPageContainer extends React.Component {
-  constructor(props){
-    super(props);
-  }
+const AboutPageContainer = () => ensureCurrentUser((currentUser) => (
+  <div className="page-wrapper d-flex flex-column">
+    <HeaderContainer current_user={currentUser}></HeaderContainer>
+    <main id="wrap" className="app-container container flex-fill">
+      <AboutPage/>
+    </main>
+    <FooterContainer></FooterContainer>
+  </div>
+));
 
-  render() {
-    return(
-      <div className="page-wrapper d-flex flex-column">
-        <HeaderContainer></HeaderContainer>
-        <main id="wrap" className="app-container container flex-fill">
-          <AboutPage/>
-        </main>
-        <FooterContainer></FooterContainer>
-      </div>
-    )
-  }
-}
+export default AboutPageContainer;
